renamed files to follow python convention (no uppercase)
