+++ manual_amp_ui.py
