+++ semi_auto_tt_ui.py
